Made RealLEDs more generic. Not tested.
