+++ src/main/deploy/robotConfigs/TopGun/LED.conf
@@ -1,5 +1,3 @@
-TopGunLEDs{
-    Red = 7
-    Green = 8
-    Blue = 9
-}
+Red = 7
+Green = 8
+Blue = 9
