+++ src/main/deploy/robotConfigs/TopGun/drivetrain.conf
@@ -57,28 +57,6 @@ teleop {
     kdelay=25
     kproportion=0.025
 }
-parkingbrake {
-    leftbrakeservomotor{
-        port = 8
-        maxPWMpulseWidthMS = 2.4
-        deadbandMax = 0.0
-        centerPWMpusleWithMS = 0.0
-        deadbandMin = 0.0
-        minPWMpulseWidthMS = 0.6
-    }
-    rightbrakeservomotor{
-        port = 9
-        maxPWMpulseWidthMS = 2.4
-        deadbandMax = 0.0
-        centerPWMpusleWithMS = 0.0
-        deadbandMin = 0.0
-        minPWMpulseWidthMS = 0.6
-    }
-    leftbrakeengaged=0.75
-    rightbrakeengaged=0.75
-    leftbrakedisengage=0.45
-    rightbrakedisengage=0.45
-}
 
 rightSideInverted=false
 ticksPerInch=29.4
