+++ src/main/java/frc/robot/subsystems/ledlights/RealLEDs.java
@@ -3,13 +3,13 @@ package frc.robot.subsystems.ledlights;
 import com.typesafe.config.Config;
 
 import edu.wpi.first.wpilibj.DigitalOutput;
+import edu.wpi.first.wpilibj.DriverStation;
 import frc.robot.Config4905;
+import frc.robot.Robot;
 
 public class RealLEDs extends LEDs {
-  public RealLEDs(String led) {
-    super();
-    System.out.println("Trying to make this LED:" + led);
-    Config conf = Config4905.getConfig4905().getLEDConfig().getConfig(led);
+  public RealLEDs() {
+    Config conf = Config4905.getConfig4905().getLEDConfig();
 
     m_red = new DigitalOutput(conf.getInt("Red"));
     m_red.enablePWM(0);
@@ -21,5 +21,50 @@ public class RealLEDs extends LEDs {
     setSolid();
 
     m_readyForPeriodic = true;
+  }
+
+  @Override
+  public void periodic() {
+    if (Robot.getInstance().isDisabled()) {
+      setRainbow();
+    } else if (Robot.getInstance().isTeleop()) {
+      double matchTime = DriverStation.getMatchTime();
+      if (matchTime <= 30 && matchTime > 0) {
+        setYellow(1);
+        setBlinking(1);
+      } else {
+        switch (getTeleOpMode()) {
+        case SLOW:
+          setBlue(1);
+          setSolid();
+          break;
+
+        case MID:
+          setPurple(1);
+          setSolid();
+          break;
+
+        case FAST:
+          setGreen(1);
+          setSolid();
+          break;
+
+        case BRAKED:
+          setRed(1);
+          setBlinking(0.05);
+          break;
+
+        default:
+          setOrange(1);
+          setSolid();
+        }
+      }
+    } else if (Robot.getInstance().isAutonomous()) {
+      setWhite(1);
+      setSolid();
+    } else {
+      setRainbow();
+    }
+    super.periodic();
   }
 }
