--- src/main/java/frc/robot/subsystems/ledlights/TopGunLEDs.java
@@ -1,67 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.subsystems.ledlights;
-
-import edu.wpi.first.wpilibj.DriverStation;
-import frc.robot.Robot;
-
-/** Add your docs here. */
-public class TopGunLEDs extends RealLEDs {
-
-  public TopGunLEDs() {
-    super("TopGunLEDs");
-  }
-
-  @Override
-  public void periodic() {
-    if (Robot.getInstance().isDisabled()) {
-      setRainbow();
-    } else if (Robot.getInstance().isTeleop()) {
-      double matchTime = DriverStation.getMatchTime();
-      if (matchTime <= 30 && matchTime > 20) {
-        setOrange(1);
-        setBlinking(1);
-      } else if (matchTime <= 20 && matchTime > 10) {
-        setYellow(1);
-        setBlinking(0.5);
-      } else if (matchTime <= 10 && matchTime >= 0) {
-        setWhite(1);
-        setBlinking(0.25);
-      } else {
-        switch (getTeleOpMode()) {
-        case SLOW:
-          setBlue(1);
-          setSolid();
-          break;
-
-        case MID:
-          setPurple(1);
-          setSolid();
-          break;
-
-        case FAST:
-          setGreen(1);
-          setSolid();
-          break;
-
-        case BRAKED:
-          setRed(1);
-          setBlinking(0.05);
-          break;
-
-        default:
-          setOrange(1);
-          setSolid();
-        }
-      }
-    } else if (Robot.getInstance().isAutonomous()) {
-      setWhite(1);
-      setSolid();
-    } else {
-      setRainbow();
-    }
-    super.periodic();
-  }
-}
